% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R, R/etl_transform.R
\name{etl_extract.etl_citibike}
\alias{etl_extract.etl_citibike}
\alias{etl_load.etl_citibike}
\alias{etl_transform.etl_citibike}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_citibike}(obj, ...)

\method{etl_load}{etl_citibike}(obj, year = 2015, month = 1, ...)

\method{etl_transform}{etl_citibike}(obj, year = 2015, month = 1, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}

\item{year}{a year represented as a four-digit integer}

\item{months}{a vector of integers representing the months}
}
\description{
etl_extract
}
\details{
If a \code{year} and/or \code{month} is specified, then
only flight data from matching months is used.
}
\examples{
bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract()

bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()

#' # check the results
\dontrun{
flights.db <- tbl(db, "flights")
flights.db \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}

