% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R, R/etl_transform.R
\name{etl_extract.etl_citibike}
\alias{etl_extract.etl_citibike}
\alias{etl_load.etl_citibike}
\alias{etl_transform.etl_citibike}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_citibike}(obj, years = 2013, months = 7, ...)

\method{etl_load}{etl_citibike}(obj, years = 2013, months = 7, ...)

\method{etl_transform}{etl_citibike}(obj, years = 2013, months = 7, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{years}{a numeric vector indicating the years to be downloaded (default is 2013)}

\item{months}{a numeric vector indicating the months to be downloaded (default is 7)}

\item{...}{arguments passed to methods}
}
\description{
etl_extract

etl_load

etl_transform
}
\details{
This function downloads NYC CitiBike data from citibike website for years and months specified.

This function loads NYC CitiBike data into a SQLite database for years and months specified.

This function unzips NYC CitiBike data for years and months specified.
}
\examples{
\dontrun{
bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()
}
}
