% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R, R/etl_transform.R
\name{etl_extract.etl_citibike}
\alias{etl_extract.etl_citibike}
\alias{etl_load.etl_citibike}
\alias{etl_transform.etl_citibike}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_citibike}(obj, years = 2015, months = 1:12, ...)

\method{etl_load}{etl_citibike}(obj, schema = FALSE, years = 2015,
  months = 1:12, ...)

\method{etl_transform}{etl_citibike}(obj, years = 2015, months = 1:12, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{years}{a numeric vector indicating the years to be downloaded (default is 2015)}

\item{months}{a numeric vector indicating the months to be downloaded (default is 1:12)}

\item{...}{arguments passed to methods}

\item{schema}{Either a logical, a filename pointing to SQL schema file, or a character
vector containing the schema itself. If \code{schema = TRUE}, then the built-in
schema will be used. Note
that the flavor of SQL in this file must match the type of the source. That is,
if your object is of type \code{\link[dplyr]{src_mysql}}, then make sure that
the schema you specify here is written in MySQL (and not PostgreSQL). Please
note that SQL syntax is not, in general, completely portable. Use with caution, as this may
clobber any existing data you have in an existing database.}
}
\description{
etl_extract

etl_load

etl_transform
}
\details{
This function downloads NYC CitiBike data from citibike website for years and months specified.

This function loads NYC CitiBike data into a SQLite database for years and months specified.

This function unzips NYC CitiBike data for years and months specified.
}
\examples{
\dontrun{
bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract(years = 2014, months = 8:9)
}
\dontrun{
bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()

trips <- tbl(bikes,"trips")
head(trips)
}
#' # check the results
\dontrun{
trips \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
\dontrun{
bikes <- etl("citibike", dir = "~/Desktop/citibike_data")
bikes \%>\%
  etl_extract(years = 2014, months = 8:9) \%>\%
  etl_transform()
}
}

